refactor: colocate achievement progress service tests
