+++ lib/__tests__/achievement-progress-service.evaluators-gold-reward.test.ts
@@ -1,5 +1,5 @@
 import { AchievementProgressService } from "../achievement-progress-service";
-import type { MockChain } from "./achievement-progress-service.fixtures";
+import type { MockChain } from "../achievement-progress-service.fixtures";
 import {
   makeDataResult,
   makeUpsertResult,
@@ -7,7 +7,7 @@ import {
   CHARACTER_ID,
   USER_ID,
   ACHIEVEMENT_ID,
-} from "./achievement-progress-service.fixtures";
+} from "../achievement-progress-service.fixtures";
 
 const mockWriteClient = {
   from: jest.fn(),
